+++ R/allGenerics.R
@@ -1,0 +1,44 @@
+#' @rdname classification
+
+
+setGeneric('classification', function(object) {
+  standardGeneric('classification')
+})
+
+#' @rdname meta
+
+setGeneric('meta', function(object) {
+  standardGeneric('meta')
+})
+
+#' @rdname description
+
+setGeneric('description', function(object) {
+  standardGeneric('description')
+})
+
+
+#' @rdname chebi
+
+setGeneric('chebi', function(object) {
+  standardGeneric('chebi')
+})
+
+#' @rdname descriptors
+
+setGeneric('descriptors', function(object) {
+  standardGeneric('descriptors')
+})
+
+
+#' @rdname alternative_parents
+
+setGeneric('alternative_parents', function(object) {
+  standardGeneric('alternative_parents')
+})
+
+#' @rdname direct_parent
+
+setGeneric('direct_parent', function(object) {
+  standardGeneric('direct_parent')
+})